refactor(desktop): split the 2,266-line volume-copy test file into eight suites named after what they cover
`transfer/volume/copy_tests.rs` was the app crate's longest file and the one a reader had to scroll to find anything in. It is now `copy_tests/` with a `mod.rs` holding the shared imports and the `make_state` / `make_volumes` fixtures, plus one child per behavior under test: `cancellation`, `conflicts`, `destination`, `error_mapping`, `progress`, `remove_tree`, `scan`, and `scan_hints` (112 to 588 lines each, all under the 800-line warn threshold).

A pure move: all 51 tests kept their names and bodies byte for byte, and 51 run and pass after the split. Only the module path changed, so the sibling `#[path]` suites and the two doc comments naming the old file were updated to match.

Verified: `cargo check --workspace --exclude cmdr-fsevent-stream --features cmdr/virtual-mtp --all-targets` clean, `cargo nextest run -E 'test(/volume::copy::tests::/)'` 51 passed (51 before).
