+++ apps/desktop/src-tauri/src/file_system/write_operations/transfer/DETAILS.md
@@ -68,7 +68,7 @@ facts that none of those carry live here:
 **Cross-type overwrites (file↔folder).** Both copy and move route Overwrite-with-type-mismatch through `overwrite::safe_overwrite_dir`:
 - Local copy `copy_single_item`'s parent-creation site: when the source tree wants a directory at a path holding a file (folder→file overwrite), the closure does `create_dir_all` while the helper sets the file aside as `<name>.cmdr-temp-<uuid>`. The symlink branch's file→folder overwrite goes through the same helper.
 - Local move `move_with_rename` / `merge_move_directory`: when `resolve_conflict` returns Overwrite for a type-mismatched pair, the closure does `fs::rename(source, target)`.
-- Volume copy/move via `volume::conflict::apply_volume_conflict_resolution`: a type swap can't temp-rename across backends, so **cross-type** Overwrite deletes the dest first (`remove_tree(…, UserChoseOverwriteAcrossTypes)` for folder dests, `Volume::delete` for file dests) before the streaming writer / recursive copy lands the source. **File→file** Overwrite does NOT delete first — it uses the safe-replace temp+finalize path (see "Cross-volume file→file Overwrite is a safe-replace" below). Same-type dir-vs-dir still skips the delete to honor the merge-not-replace guarantee. Pinned by `volume/copy_tests.rs::test_volume_overwrite_{file_over_existing_folder,folder_over_existing_file}`.
+- Volume copy/move via `volume::conflict::apply_volume_conflict_resolution`: a type swap can't temp-rename across backends, so **cross-type** Overwrite deletes the dest first (`remove_tree(…, UserChoseOverwriteAcrossTypes)` for folder dests, `Volume::delete` for file dests) before the streaming writer / recursive copy lands the source. **File→file** Overwrite does NOT delete first — it uses the safe-replace temp+finalize path (see "Cross-volume file→file Overwrite is a safe-replace" below). Same-type dir-vs-dir still skips the delete to honor the merge-not-replace guarantee. Pinned by `volume/copy_tests/conflicts.rs::test_volume_overwrite_{file_over_existing_folder,folder_over_existing_file}`.
 
 **Copy strategy selection** (`copy_strategy.rs`). `select_local_copy_strategy` names the mechanism, `copy_file_using` runs it:
 - macOS, same APFS volume → `LocalCopyStrategy::AppleClone`, `copyfile(3)` with `COPYFILE_CLONE` for instant clonefile
