+++ apps/desktop/src-tauri/src/file_system/write_operations/transfer/volume/copy_rollback_tests.rs
@@ -1,12 +1,12 @@
 //! Rollback / dest-inside-source / cancel-mid-stream tests for `volume::copy`,
-//! split out of `volume/copy_tests.rs`. These cover user-initiated Rollback at
+//! split out of `volume/copy_tests/`. These cover user-initiated Rollback at
 //! various points (after the first file, mid-merge-stream, after a finalize
 //! rename), the rejection of copying a directory into its own descendant, and
 //! cancel-mid-stream data preservation. The `RollbackAfterFirstFileSink`,
 //! `TripIntentOnFirstByteSink`, and `TripIntentAtFilesDoneSink` doubles trip the
 //! operation's intent at the precise moment each scenario needs.
 //!
-//! Shared fixtures `make_state` / `make_volumes` live in `volume/copy_tests.rs`
+//! Shared fixtures `make_state` / `make_volumes` live in `volume/copy_tests/mod.rs`
 //! (`super::tests`) so they aren't duplicated.
 
 use super::tests::{make_state, make_volumes};
